Move base testcase to the common namespace.
This way, it'll be more logical to also use it to test Fork
"controllers" too.
