--- src/ForkCMS/Bundle/InstallerBundle/Test/WebTestCase.php
@@ -1,43 +1,0 @@
-<?php
-
-namespace ForkCMS\Bundle\InstallerBundle\Test;
-
-use Symfony\Bundle\FrameworkBundle\Test\WebTestCase as BaseWebTestCase;
-use Symfony\Component\Finder\Finder;
-
-/**
- * WebTestCase is the base class for functional tests.
- *
- * @todo Remove this when Fork has no custom Kernel class anymore
- * @author Wouter Sioen <wouter.sioen@gmail.com>
- */
-abstract class WebTestCase extends BaseWebTestCase
-{
-    /**
-     * Attempts to guess the kernel location.
-     *
-     * When the Kernel is located, the file is required.
-     *
-     * @return string The Kernel class name
-     *
-     * @throws \RuntimeException
-     */
-    protected static function getKernelClass()
-    {
-        $dir = isset($_SERVER['KERNEL_DIR']) ? $_SERVER['KERNEL_DIR'] : static::getPhpUnitXmlDir();
-
-        $finder = new Finder();
-        $finder->name('AppKernel.php')->depth(0)->in($dir);
-        $results = iterator_to_array($finder);
-        if (!count($results)) {
-            throw new \RuntimeException('Either set KERNEL_DIR in your phpunit.xml according to http://symfony.com/doc/current/book/testing.html#your-first-functional-test or override the WebTestCase::createKernel() method.');
-        }
-
-        $file = current($results);
-        $class = $file->getBasename('.php');
-
-        require_once $file;
-
-        return $class;
-    }
-}
